WIP SCM similartiy must be in [0.01, 1.0]

+++ project/src/main/java/es/ubu/lsi/avrela/css/adapter/web/WebScmCaseStudySimulation.java
@@ -15,7 +15,7 @@ public class WebScmCaseStudySimulation {
 
   private WebCommitSimilarityFunctionConfig commitSimilarityFunctionConfig;
 
-  private Integer similarityThreshold;
+  private Double similarityThreshold;
 
   private ScmWebRubricEvaluation  rubricEvaluation;
